refactor: move oauth package
