+++ src/main/kotlin/com/relogging/server/global/oauth/handler/OAuthAuthenticationSuccessHandler.kt
@@ -1,4 +1,4 @@
-package com.relogging.server.oauth.handler
+package com.relogging.server.global.oauth.handler
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
